# public/swagger/json/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "UNOARITMETICO",
        "description": "API unoaritmetico, to use endpoints that need authentication use the 'Set Globally' function first",
        "contact": {
            "email": "walook.com.mx"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "unoaritmetico.com",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/groups": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Obtains groups",
                "description": "Obtains groups",
                "operationId": "getGroups",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page requested, default is 1",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page, default is 10",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Groups created by user",
                        "schema": {
                            "$ref": "#/definitions/GetGroupsResponse"
                        }
                    },
                    "400": {
                        "description": "Request format isn't valid",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Token is invalid",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Invalid Method",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Create a group",
                "description": "Create a group",
                "operationId": "storeGroup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Account to register",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/StoreGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account was registered",
                        "schema": {
                            "$ref": "#/definitions/StoreGroupResponse"
                        }
                    },
                    "405": {
                        "description": "Invalid Method",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "422": {
                        "description": "Invalid Fields",
                        "schema": {
                            "$ref": "#/definitions/Validation"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/groups/{groupId}": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Display a single group",
                "description": "Display a single group",
                "operationId": "getGroups",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "Id of group to retrieve",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Groups created by user",
                        "schema": {
                            "$ref": "#/definitions/DetailGroupResponse"
                        }
                    },
                    "400": {
                        "description": "Request format isn't valid",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Token is invalid",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Invalid Method",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Groups"
                ],
                "summary": "Update a group",
                "description": "Update a group",
                "operationId": "updateGroup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Group to update",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateGroupRequest"
                        }
                    },
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "Id of group to retrieve",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group was updated",
                        "schema": {
                            "$ref": "#/definitions/UpdateGroupResponse"
                        }
                    },
                    "405": {
                        "description": "Invalid Method",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "422": {
                        "description": "Invalid Fields",
                        "schema": {
                            "$ref": "#/definitions/Validation"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Groups"
                ],
                "summary": "Delete a group",
                "description": "Delete a group",
                "operationId": "deleteGroup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "Id of group to delete",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Group was deleted"
                    },
                    "405": {
                        "description": "Invalid Method",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "422": {
                        "description": "Invalid Fields",
                        "schema": {
                            "$ref": "#/definitions/Validation"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Obtains all students",
                "description": "Obtains all students",
                "operationId": "getStudents",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page requested, default is 1",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page, default is 10",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Students created by user in the group",
                        "schema": {
                            "$ref": "#/definitions/GetStudentsResponse"
                        }
                    },
                    "400": {
                        "description": "Request format isn't valid",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Token is invalid",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Invalid Method",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Create a student",
                "description": "Create a student",
                "operationId": "storeStudent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Student information",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/StoreStudentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account was registered",
                        "schema": {
                            "$ref": "#/definitions/StoreStudentResponse"
                        }
                    },
                    "405": {
                        "description": "Invalid Method",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "422": {
                        "description": "Invalid Fields",
                        "schema": {
                            "$ref": "#/definitions/Validation"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/students/{studentId}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Display a single student",
                "description": "Display a single student",
                "operationId": "getGroups",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "studentId",
                        "in": "path",
                        "description": "Id of student to retrieve",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Groups created by user",
                        "schema": {
                            "$ref": "#/definitions/DetailStudentResponse"
                        }
                    },
                    "400": {
                        "description": "Request format isn't valid",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Token is invalid",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Invalid Method",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Update a student",
                "description": "Update a student",
                "operationId": "updateStudent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Group to update",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateStudentRequest"
                        }
                    },
                    {
                        "name": "studentId",
                        "in": "path",
                        "description": "Id of student to update",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group was updated",
                        "schema": {
                            "$ref": "#/definitions/UpdateStudentResponse"
                        }
                    },
                    "405": {
                        "description": "Invalid Method",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "422": {
                        "description": "Invalid Fields",
                        "schema": {
                            "$ref": "#/definitions/Validation"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Students"
                ],
                "summary": "Delete a student",
                "description": "Delete a student",
                "operationId": "deleteStudent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "studentId",
                        "in": "path",
                        "description": "Id of student to delete",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Group was deleted"
                    },
                    "405": {
                        "description": "Invalid Method",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "422": {
                        "description": "Invalid Fields",
                        "schema": {
                            "$ref": "#/definitions/Validation"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "StoreGroupRequest": {
            "properties": {
                "name": {
                    "description": "",
                    "type": "string",
                    "example": "Primary 1st Matutine"
                }
            }
        },
        "UpdateGroupRequest": {
            "properties": {
                "name": {
                    "description": "",
                    "type": "string",
                    "example": "Primary 1st Matutine"
                }
            }
        },
        "StoreStudentRequest": {
            "properties": {
                "name": {
                    "description": "",
                    "type": "string",
                    "example": "Mariana"
                },
                "age": {
                    "description": "",
                    "type": "integer",
                    "example": 16
                },
                "group_id": {
                    "description": "",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "UpdateStudentRequest": {
            "properties": {
                "name": {
                    "description": "",
                    "type": "string",
                    "example": "Mariana"
                },
                "age": {
                    "description": "",
                    "type": "integer",
                    "example": 16
                },
                "group_id": {
                    "description": "",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "Error": {
            "properties": {
                "message": {
                    "description": "",
                    "type": "string",
                    "example": "Bad Request"
                },
                "status_code": {
                    "description": "",
                    "type": "integer",
                    "format": "int64",
                    "example": 400
                }
            }
        },
        "Validation": {
            "properties": {
                "error": {
                    "description": "",
                    "type": "array",
                    "$ref": "#/definitions/ValidationItem"
                }
            }
        },
        "ValidationItem": {
            "properties": {
                "field": {
                    "description": "",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "DetailGroupResponse": {
            "properties": {
                "meta": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/DefaultMeta"
                },
                "data": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/GroupItem"
                }
            }
        },
        "GetGroupsResponse": {
            "properties": {
                "meta": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/PaginationMeta"
                },
                "data": {
                    "description": "",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/GroupItem"
                    }
                }
            }
        },
        "StoreGroupResponse": {
            "properties": {
                "meta": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/DefaultMeta"
                },
                "data": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/GroupItem"
                }
            }
        },
        "UpdateGroupResponse": {
            "properties": {
                "meta": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/DefaultMeta"
                },
                "data": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/GroupItem"
                }
            }
        },
        "DefaultMeta": {},
...
        "DetailStudentResponse": {
            "properties": {
                "meta": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/DefaultMeta"
                },
                "data": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/StudentItem"
                }
            }
        },
        "GetStudentsResponse": {
            "properties": {
                "meta": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/PaginationMeta"
                },
                "data": {
                    "description": "",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/StudentItem"
                    }
                }
            }
        },
        "StoreStudentResponse": {
            "properties": {
                "meta": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/DefaultMeta"
                },
                "data": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/StudentItem"
                }
            }
        },
        "UpdateStudentResponse": {
            "properties": {
                "meta": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/DefaultMeta"
                },
                "data": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/StudentItem"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiAuth": {
            "type": "oauth2",
            "name": "Authorization",
            "flow": "password",
            "tokenUrl": "unoaritmetico.com/api",
            "scopes": {
                "read:users": "Consult userss",
                "write:users": "Modify users"
            }
        }
    }
}